+++ src/features/tutorApplication/ui/diplomasUpload/data.ts
