+++ src/features/tutorApplication/ui/diplomasUpload/index.module.scss
