+++ src/shared/ui/input/index.tsx
