+++ src/shared/ui/select/index.module.scss
@@ -6,7 +6,7 @@
   &__label {
     font-size: 20px;
     font-weight: 700;
-line-height: 1;
+    line-height: 1;
   }
 }
